// newsportal_backend/src/app/role/role.interface.ts
export interface IRoleInterface {
  _id?: any;
  role_name: string;
  dashboard_show?: true | false;
  category_show?: true | false;
  category_create?: true | false;
  category_update?: true | false;
  category_delete?: true | false;
  sub_category_show?: true | false;
  sub_category_create?: true | false;
  sub_category_update?: true | false;
  sub_category_delete?: true | false;
  today_news_category_show?: true | false;
  today_news_category_create?: true | false;
  today_news_category_update?: true | false;
  today_news_category_delete?: true | false;
  social_media_create?: true | false;
  social_media_update?: true | false;
  social_media_delete?: true | false;
  ads_show?: true | false;
  ads_create?: true | false;
  ads_update?: true | false;
  ads_delete?: true | false;
  news_show?: true | false;
  news_create?: true | false;
  news_update?: true | false;
  news_delete?: true | false;
  staff_show?: true | false;
  staff_create?: true | false;
  staff_update?: true | false;
  staff_delete?: true | false;
  online_jorip_show?: true | false;
  online_jorip_create?: true | false;
  online_jorip_update?: true | false;
  online_jorip_delete?: true | false;
  staff_permission_show?: true | false;
  staff_permission_create?: true | false;
  staff_permission_update?: true | false;
  staff_permission_delete?: true | false;
  site_setting_update?: true | false;
}